kinetics: Added alpha-QSS method for chemical kinetics update.
Also, during testing found a bug in the contolling timestep
algorithm that allowed successively smaller timesteps to
be saved and used on the next update. The code could work itself
into taking exceedingly small timesteps. This bug is fixed by
using a dtSave variable to record the last "good" timestep.