gas/nm/util: Make list of files local to directories...
I was starting to get annoyed with maintaining multiple
lists of files to include when building D executables in
various areas. In this commit, I've started to house a file
called $DIR_files.mk in each of the separate areas of the source.
The higher-level code requiring those files can then use
"include" directives in their makefiles to get access to the
list of files.

Also, I've started naming modules consistently with a package
designator in front. I changed a few modules in "nm" but this
has probably broken the build in a number of spots. The build
works for the demo programs in the gas/ area.