kinetics: Added an RKFStep unit test.
The unit test checks that the order of accuracy for
the RKFStep is in fact 5th order. It does this by
solving a chemical kinetics problem over the same
interval twice, with a different timestep on each attempt.
We can then check that the error drops as the 5th power
of the timestep reduction.