gas: Lua massf table now indexed via species names
In order to add this, I needed a reverse lookup method
available in the GasModel that could give me the index
of a species given its string name. Naturally, this change
needed to propagate to all derived models. Hence, the
large number of files changed in this commit.

I'm pleased with the unit tests though. It gives me a fair
amount of confidence that I haven't broken anything.