geom: Small and subtle fix for the Vector3 normalize method for DMD 2.069.0.
There's probably something that I don't understand about opAssign but, for
DMD 2.069.0, we couldn't just divide the Vector3 object by (scalar) magnitude
to correctly get the unit vector.  This seems to be a new behaviour for DMD
but the simple fix is to operate directly on the underlying array of components.

Also, put in some lines to flush any very small components to zero.
This should tidy things up for grids that are axis-aligned but end up
with tiny nonzero components int their normal and tangential vectors
because of floating-point round-off.