gas: massf[0] = 1.0 is default for Lua-wrapped GasState when n_species = 1
To avoid some surprises for users who only ever deal with
single species gas models (and hence largely ignore the 'massf' array),
the default action is to set the value of massf[0] to 1.0 when
a single species gas is in play.

For multi-component gases, the behaviour of setting the massf
table remains the same: assume a value of 0.0 mass fraction
for all species *not* explicitly listed in the table.

I don't expect too many surprises for users of multi-component
gases since they will need to think about setting the gas composition
at some point.