eilmer: make the UBlock class known to the main code.
This is mainly using the base class Block rather than SBlock in the simulation
core and making separate places in the boundary condition classes for structured
and unstructured grid code.  A few features of the structured-grid code have
had to appear in the unstructured-grid classes in an effort to ease the process
of integration.  To mangle a number of metaphors, elegance and purity are now
out the window as we get out the bigger hammer to finish the job of forcing
the square peg of unstructured grids into the round hole of a classically
structured-grid code :)