eilmer: fixed failure of the n90 reacting case on 32-bit systems.
It seems that subtle floating-point behaviour has bitten me again.
The apparent failure of this case, only on 32-bit machines, involved a couple
of floating-point calculation issues.

First, the precision of the brent function solver was too tight for the
imprecision of the 32-bit code when computing energy differences at moderately
large temperatures (of a couple of thousand degrees).  That was easily fixed
by setting the tolerance on the energy difference to 0.01 Joules.  I do not
understand why the 32-bit code should be imprecise; it's using doubles for the
floating-point quantities.

Fixing that uncovered a second, separate problem.
For some, not-yet-understood reason, the bracketing procedure was saying
that it had finished its work but the follow-up solver was taking a different
view and complaining that the bracket really was not correct.  This difference
of view was happening only occasionally, when the left point was right on top
of the desired solution.  The current sidestep for this issue is to perturb
both ends of the temperature range, away from the incoming guess, even though
it may be very good.

Along the way, we have improved the error reporting when the relevant exceptions
are raised.