eilmer: Increase the tolerance on the number of steps for the n90 reacting case.
There are small differences in the floating-point results between some versions
of the DMD32 and DMD64 compilers.  Although I don't understand the cause of
this issue, I would prefer not to have reports of test failures because of the
small difference in the number of time-steps taken while all other measures
appear OK.  Large differences, yes, I want to know about them.