eilmer/gas: explicitly use two initialised Lua_states.
While using valgrind to track some memory problems, I want
to de-clutter the output as much as possible. Part of that
involved closing some lua_States that are opened during
eilmer simulations: one in the gas model configuration and
one related to the block-local lua_State. For the latter,
we now provide a destructor [~this()] for the Block class.