eilmer: convective flux calculation now uses the references to cells that were stored in the FVInterface object.
This makes the convective flux calculation a little be more insulated from
the underlying structured grid.  We should be able to use almost the same code
for the unstructured grid, hopefully.

Of course, we also had to store the extra left and right cell references
so that they were available to the convective-fluc function and the 1D
interpolation functions.