eilmer: Consistently expect named-field mass fraction tables from users.
I hadn't updated the interaction with the user-defined boundary conditions
to use the new approach of interacting with mass fraction tables with
named-fields only. This is now done.

After a small change to the udf-bc.lua file, the cone20-udf case is
now working again. I also removed the separate udf files: extrapolate-out
and slip-wall as these appear to not be used by anything.