eilmer: flow gradients may now be evaluated directly at faces, in 2D, at least.
Assembly of the 3D poing clouds is yet to be done.

By default, the gradients are still calculated at the cell vertices and then
averaged to get the cell face gradients.  Forcing the gradient calculations
to be done only at cell faces is activated in the user-input file with
config.deriv_calc_at_vertices = false