eilmer: Added checks that b.c.'s are of appropriate type at preparation stage.
Since we are using Lua tables to hold b.c. information at the preparation
stage, it is not possible to use the b.c's "class type" to determine if
an appropriate boundary condition has been set. This meant it was possible
to set a solid domain b.c. at a gas domain boundary, and vice versa.

To remedy this, I've added boolean flags to each of the gas and solid domain
b.c. tables. These are checked when the config file is written to ensure that
a b.c. appropriate to the domain type is being set. If it's not appropriate,
an error is issued and the preparation program terminates.