geom: Added joinGrid method for StructuredGrid objects.
Presently, this joinGrid method only works in 2D and
assumes the grids to be joined have the same orientation.
It can handle:
  1. east-west joins
  2. north-south joins

One of the envisaged use cases for this method is to
rejoin a grid after it might have been split by
the subgrid method (such as what happens in SBlockArray).
A series of applications of joinGrid can be used to
rebuild a single grid.