eilmer/examples: Completed README for lobb-sphere example.
While completing the README, I was annoyed with how clunky my
approach was to finding the final time index. It was difficult
to document, so I removed the difficulty altogether by providing
a tindx='last' option to the FlowSolution constructor. I updated
the example to use this option in the shock-detachment.lua.

Also, caught one more auto-generated file I had added in the smoke-test
area and removed that with this commit.