eilmer: Add checks that the b.c. constructor is called by user.
This avoids a possible point of confusion for user's script.
If one were to set a bcList like:

  bcList={north=OutFlowBC_Simple, east=WallBC_WithSlip}

they would slip through the preparation program without complaint,
but the b.c. would not be set as intended. In fact, it would not
be set at all. The user must always call the constructor ":new{}"
so that some behind-the-scenes configuration can take place.
The correct way to do the above is:

  bcList={north=OutFlowBC_Simple:new{}, east=WallBC_WithSlip:new{}}

The code will now detect the case of an improperly configured boundary
condition. It cannot detect this at declaration but only when it
goes to process the boundary condition. The preparation program will
issue an error message and exit if the boundary condition is not
configured. We detect this with an "is_configured" boolean as part
of the BoundaryCondition class.