eilmer: Do the reduction to find the minimum dt_allow in a serial loop.
This is to avoid a race condition when applying the min function and then
updating the global variable within a parallel loop.  By the time any particular
block acquires the lock and updates the global variable dt_allow, the value
upon which it did made its decision may be stale.