Eilmer: added code for high-order reconstruction for unstructured-grid blocks.
It doesn't quite work and the compile time has blown out to nearly 8 minutes on
my 32-bit Linux Mint system and DMD32 D Compiler v2.070.0, probably because of
the use of more mixins.
The doesn't quite work bit is about the calculation going awry with a matrix
that it claims is effectively singular but the debug for that particular matrix
doesn't seem to appear on the console.  This commit allows me to move to a
64-bit machine to see if I can dig out any more details.