kinetics: Two changes to chemistry timestep selection.
I've made what appear to be two subtle changes to the chemistry
timestep selection. They might turn out to be profound. Some
rigorous testing over the next few days will tell me if I've caused
problems.

Change 1: We'd like to store what happens on the penultimate step
of the cycling since the ultimate step might be only a small fraction
in size. We were storing the actual size of the penultimate step as
our new suggested timestep. It might in fact be better to store
the penultimate suggested timestep for starting subcycling on our
next chemistry update. That's what we do now. Keep a hold of the
suggested timestep.

Change 2: If we only take a single step, that's a special case:
we have no penultimate step to hold on to. In that case, store
the suggested timestep if we are ready to leave the routine.