Eilmer: a very satisfactory debugging session for the 3D unstructured-grid.
The key problem with the slow flow development of the simple-ramp-usg case
was found on line 428 of ublock.d.  I wasn't passing corner points correctly.
The result was the quad properties were for one of their (half) triangles.

There was another more subtle problem with the out_sign values for north and
south cell interfaces in geom/usgrid.d.