eilmer: Make the configuration of the gasdynamic update aware of moving grids.
To do this, we introduce two specific moving-grid update scheme names:
"moving_grid_1_stage" and "moving_grid_2_stage", one of which must be selected
in order to have a calculation with moving grid.  The details of the grid motion
remain a separate concern.

Along the way, we fixed a small bug for verbose output, now that boundary
conditions for the blocks are constructed a little later than they used to be.
And, abstracted the final-index calculation to s single function that is called
in a small number of places.