nm: Added dot product functions that will work on a restricted range of a matrix.
In the GMRES method, the working matrices grow with each step of the iteration.
It seems more attractive from a memory allocation standpoint, to pre-allocate
a maximum anticipated matrix size and then only work with a restricted subset
of those pre-allocated matrices during matrix operations. The new dot product
functions provide that capability to work on a restricted portion of the
upper left part of the supplied matrices.