nm: Added a more efficient version of the GMRES iterative solver.
The function 'gmres2' is more efficient than the first version, 'gmres',
because it will terminate the iterations as soon as the required drop
in residual is achieved. The first version will iterate for a fixed number
of iterations only or if it finds the exact answer.

In order to compute the residual on each step, the rotated Hessenberg
matrix needs to be built-up progressively on each iteration. This adds
some complexity to the algorithm. I have followed the hints given on
p.178 of Saad's text to implement this approach.

   Saad (2003)
   Iterative Methods for Sparse Linear Systems, 2nd ed.
   SIAM, Philadelphia