eilmer: First additions for steady-state solver mode.
In this commit, I've added some of the pieces required to
compute a low-order Jacobian via finite-differences. This
code is not hooked up or thoroughly tested yet.

The steady-state solver will be available when compiled with
-version=steadystate. This is because the steady-state solver
will be rather more memory hungry than a time-accurate solve
with explicit updates. The users of the time-accurate solver
should not need to deal with this memory overhead.