kinetics/eilmer: Refactoring to move working arrays into ChemistryUpdate class.
I renamed ReactionUpdateScheme to ChemistryUpdate. I also made the action
of ChemistryUpdate class callable (via opCall) because I don't like
classes with methods like "doIt()". This had a small ripple effect
through the eilmer calling code and has been updated as part of this commit.

I also introduced a ChemistryUpdateException that is thrown if the
chemistry update fails for any reason. The 'message' string of the
the ChemistryUpdateException will give the reason for the failure.