eilmer: Throw away low-order Jacobian code and prepare for nested GMRES solvers.
I hit a dead-end with using a low-order Jacobian, decomposed with the ILU(0)
method: my matrix for compressible Euler flow is not diagonally dominant and
contains small and zero pivots. After much reading, the solutions to this seemed
overly complex and not very robust, for all their complexity. So, I threw away
a lot of code today related to computing the low-order Jacobian and storing
it in a sparse matrix. Ken Thompson would be proud (see quote)

   One of my most productive days was throwing away 1000 lines of code.
                                                -- Ken Thompson

Instead, I've gone with the nested GMRES methods. On the theory that
one GMRES iterative solver is good, then two must be better, I intend
to use Saad's Flexible GMRES algorithm that allows for a variable
preconditioner. The preconditioner itself will be a GMRES iterative solve.
The changes to block.d are to accommodate this new approach.

I think this method will greatly simplify implicit boundary conditions too.