eilmer: Changed user-facing names of UDF b.c.'s
There are now just two functions that the user should supply
in their Lua scripts for user-defined boundary conditions:

function ghostCell(args) ... end
function interface(args) ... end

The function will be supplied with a boundaryId arg so that
the user may specialise the treatment based on which boundary
the function is being applied at. For example, to do something
special at boundary 2:

   function ghostCell(args)
      if args.boundaryId == 2 then
         """Do something special at 2"""
         return ghostCell0, ghostCell1
      end
   end

For structured grids, the user is also given global names
'north', 'east', 'south' and 'west' to test against so
that explicit boundary numbers aren't neded. For example, the
delegator function in cone20-udf works in this way:

   function ghostCell(args)
      -- Delegate to the appropriate sub-function
      if args.boundaryId == west then
         return ghostCell_west(args)
      elseif args.boundaryId == east then
         return ghostCell_east(args)
      else
         print("No ghost cell b.c. function available on boundary: ", args.boundaryId)
      end
   end

This change unifies the udf treatment of ghost cells and interfaces for
both structured and unstructured grids.

I have also update some of the examples to reflect this change. I should look
for more examples that have been affected.