eilmer/examples: Simplified udf-bc.lua in MMS example to use ghost-cell positions directly.
We had been cheating a little with the sampleFlow function to probe the positions for
the ghost cells beyond the edge of the domain. This approach did not generalise to
unstructured grids. Instead, we can make use of the "gc0{x|y|z|}" and "gc1{x|y|z}"
values in the supplied "args" table. PJ had made these available to do the vortex
test case. This greatly simplifies the UDF file. It removes all of the specialisation
required for different boundaries of a structured grid. In fact, the udf-bc.lua files
should now work for both the structured and unstructured grids cases.