eilmer: prepared to include one more cell in the lists either side of a face.
Kyle pointed out that my workbook showed left- and right-lists of points that
included the cells either side of the face.  So far the code has set up the
lists such that they do not include the one cell on the other side of the face.
This commit prepares the code to be able to include those cells when we expect
that they will have valid flow data.

Currently, when we do include those points the high-order reconstruction appears
to go badly and we end up with a bad cell for the cone20 and sod USG examples.
Some debugging is in order, so we have added some extra output for the toString
methods of the cells and faces.

The present commit leaves the code in a known good state, in that the test cases
are running successfully as I type this log message.  Up to bittker so far.