eilmer: Point clouds now can include an extra point for unstructured         reconstruction. Fixed bug in least-squares interpolation.
Point clouds now may have up to 5 points, the extra point is the
neighbour cell on the opposite side of the interface (i.e. for a
left_cells cloud the 5th point is the centre point of the right_cells
cloud). Also, boundary clouds now include a third internal domain cell
(in addition to the original two ghost cells).

NB: now that the left_cells and right_cells clouds each have
the possibility of sharing points, we should be careful not to
transform the interface neighbour cells twice!

A bug was also fixed in the least squares reconstruction where the
reconstructed values (qL and qR) were not being set to the Lft
and Rght flowstates.

G: Enter commit message.  Lines beginning with 'HG:' are removed.