eilmer: Added in code to fill out cloud_pos and cloud_fs used in unstructured         grid spatial derivatives calculations for viscous simulations.
It has been decided to currently only offer support for least-squares
spatial derivative calculations for the unstructured grid using data
stored at interface mid-points. This update pushes the code which loops
through the boundary interfaces and internal interfaces and stores the
cloud of data needed for the spatial derivative calculation.

Currently there is a chunk of repeated code for the internal interfaces
and boundary interfaces. In the future we will refactor this code to
keep in line with the DRY principle.

This update also sets the default setting of deriv_calc_at_vertices to
false, this ensures that both structured and unstructured grid viscous
simulations run straight out of the box.