eilmer: First hack at a using finite-differences to compute         the gradients for viscous transport. Added back in         the symmetric least-squares boundary cloud.
This update mainly adds the first hack at using finite-differences
to calculate the gradients at cell interfaces used in the viscous
transport update. Currently there is a bug in either the implementation
or our understanding of the problem. More details are provided in
the code, in brief, it is currently broken and should not be used.

This update also returns the symmetric boundary cloud used in the
least-squares gradient calculation for viscous transport. This cloud provides
2nd order convergence even for multi-block simulations. The user
now has an option to use this cloud (which requires ghost cell data) or
the original cloud which uses no ghost cell data.