eilmer: Naming of solid user-defined BCs now consistent with flow domain.
I've removed the solid user-defined BCs of the form:

solidInterface_north, solidInterface_south, etc.

Now, the use just specifies a single:

function solidInterface(args) .... end

and will need to specialise themselves based on which boundary
they are on. This change makes the solid domain consistent with
the flow domain. It is motivated by the change to a single Lua context
per block, instead of per boundary.