Eilmer: Added in ability to plot along a streamline in postprocessing.
This commit adds in the ability for a user to extract solution data
along a streamline. To extract data along a streamline, add
--extract-streamline="x,y,z" to the --post command.
The post-processor collects data along the streamline passing
through the point provided (i.e. the point x, y, z is a locus for the
streamline). Mutliple streamlines can be collected by separating the
the positions by a semi-colon, e.g. "x1,y1,z1;x2,y2,z2".

There is an example of extracting and plotting streamline data in
the cone20-usg case found in the examples directory.

Currently the implementation uses a fixed initial stepsize of
of dt = 1e-6 with some logic built in to ensure we step out of a
cell. In the future we may make it possible for the user to choose this
or make the initial stepsize selection more intelligent.

NB: this implementation required a small change to the
find_enlosing_cell() method in grid.d, the switch logic initially
had dimensions == 2, this was altered to switch (dimensions).